+++ tests/dummy/config/environment.js
@@ -39,8 +39,8 @@ module.exports = function (environment) {
   if (environment === 'development') {
     // if you need to test against dev API/data:
     const env = 'dev';
-    // dev item id: 015965c7e6b94a6e9b52b78ae95c0bce
-    ENV.torii.providers['arcgis-oauth-bearer'].apiKey = '2Luiwqujuz7qQwo5';
+    // dev item id: b1e048699a6943dfa9a5f1f6b325998c
+    ENV.torii.providers['arcgis-oauth-bearer'].apiKey = '7xJPH3g6bgpP8MNw';
 
     // if you need to test against QA API/data:
     // const env = 'qa';